backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
python-dotenv
openai